append_history
